moved DestinationTag to types directory
