+++ xrpl/transaction/check_create_test.go
@@ -27,7 +27,7 @@ func TestCheckCreate_Flatten(t *testing.T) {
 				},
 				Destination:    "rsA2LpzuawewSBQXkiju3YQTMzW13pAAdW",
 				SendMax:        types.XRPCurrencyAmount(10000),
-				DestinationTag: DestinationTag(23480),
+				DestinationTag: types.DestinationTag(23480),
 				Expiration:     533257958,
 				InvoiceID:      "A0258020E3B0C44298FC1C149AFBF4C8996FB92427AE41E4649B934CA495991B7852B855810100",
 			},
@@ -84,7 +84,7 @@ func TestCheckCreate_Validate(t *testing.T) {
 				},
 				Destination:    "rsA2LpzuawewSBQXkiju3YQTMzW13pAAdW",
 				SendMax:        types.XRPCurrencyAmount(10000),
-				DestinationTag: DestinationTag(23480),
+				DestinationTag: types.DestinationTag(23480),
 				Expiration:     533257958,
 				InvoiceID:      "A0258020E3B0C44298FC1C149AFBF4C8996FB92427AE41E4649B934CA495991B7852B855810100",
 			},
@@ -100,7 +100,7 @@ func TestCheckCreate_Validate(t *testing.T) {
 				},
 				Destination:    "rsA2LpzuawewSBQXkiju3YQTMzW13pAAdW",
 				SendMax:        types.XRPCurrencyAmount(10000),
-				DestinationTag: DestinationTag(23480),
+				DestinationTag: types.DestinationTag(23480),
 				Expiration:     533257958,
 				InvoiceID:      "A0258020E3B0C44298FC1C149AFBF4C8996FB92427AE41E4649B934CA495991B7852B855810100",
 			},
