+++ xrpl/transaction/payment_channel_create_test.go
@@ -32,7 +32,7 @@ func TestPaymentChannelCreate_Flatten(t *testing.T) {
 				SettleDelay:    86400,
 				PublicKey:      "32D2471DB72B27E3310F355BB33E339BF26F8392D5A93D3BC0FC3B566612DA0F0A",
 				CancelAfter:    533171558,
-				DestinationTag: DestinationTag(23480),
+				DestinationTag: types.DestinationTag(23480),
 			},
 			expected: `{
 				"Account":       "r2UeJh4HhYc5VtYc8U2YpZfQzY5Lw8kZV",
@@ -57,7 +57,7 @@ func TestPaymentChannelCreate_Flatten(t *testing.T) {
 				SettleDelay:    86400,
 				PublicKey:      "32D2471DB72B27E3310F355BB33E339BF26F8392D5A93D3BC0FC3B566612DA0F0A",
 				CancelAfter:    533171558,
-				DestinationTag: DestinationTag(0),
+				DestinationTag: types.DestinationTag(0),
 			},
 			expected: `{
 				"Account":       "r2UeJh4HhYc5VtYc8U2YpZfQzY5Lw8kZV",
@@ -123,7 +123,7 @@ func TestPaymentChannelCreate_Validate(t *testing.T) {
 				SettleDelay:    86400,
 				PublicKey:      "32D2471DB72B27E3310F355BB33E339BF26F8392D5A93D3BC0FC3B566612DA0F0A",
 				CancelAfter:    533171558,
-				DestinationTag: DestinationTag(23480),
+				DestinationTag: types.DestinationTag(23480),
 			},
 			wantValid: true,
 			wantErr:   false,
@@ -139,7 +139,7 @@ func TestPaymentChannelCreate_Validate(t *testing.T) {
 				SettleDelay:    86400,
 				PublicKey:      "32D2471DB72B27E3310F355BB33E339BF26F8392D5A93D3BC0FC3B566612DA0F0A",
 				CancelAfter:    533171558,
-				DestinationTag: DestinationTag(23480),
+				DestinationTag: types.DestinationTag(23480),
 			},
 			wantValid:   false,
 			wantErr:     true,
