+++ xrpl/transaction/types/destination_tag.go
@@ -1,4 +1,4 @@
-package transaction
+package types
 
 func DestinationTag(value uint32) *uint32 {
 	return &value
